requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.4
idna==3.11
requests==2.32.5
urllib3==2.6.3

pip~=23.2.1
python-dotenv~=1.2.2